port v14.2 handler, utils and tests to v15 handler
