+++ app/upgrades/v15/utils_test.go
@@ -1,6 +1,6 @@
 // Copyright Tharsis Labs Ltd.(Evmos)
 // SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/evmos/blob/main/LICENSE)
-package v142_test
+package v15_test
 
 import (
 	"encoding/json"
